Twitter data retrieval will now retrieve data against a certain account for up to a year back and then filter the results

+++ src/main/java/stoxology/datacollator/twitterapi/entities/Medium____.java
@@ -58,7 +58,7 @@ public class Medium____ {
     @JsonProperty("source_status_id_str")
     private String sourceStatusIdStr;
     @JsonProperty("source_user_id")
-    private Integer sourceUserId;
+    private Long sourceUserId;
     @JsonProperty("source_user_id_str")
     private String sourceUserIdStr;
     @JsonIgnore
@@ -310,7 +310,7 @@ public class Medium____ {
      *     The sourceUserId
      */
     @JsonProperty("source_user_id")
-    public Integer getSourceUserId() {
+    public Long getSourceUserId() {
         return sourceUserId;
     }
 
@@ -320,7 +320,7 @@ public class Medium____ {
      *     The source_user_id
      */
     @JsonProperty("source_user_id")
-    public void setSourceUserId(Integer sourceUserId) {
+    public void setSourceUserId(Long sourceUserId) {
         this.sourceUserId = sourceUserId;
     }
